Update install.sh. Ask for y/n for dependencies. Upgrade gcc. Use github releases to distribute gcc and nordic SDK.

+++ scripts/install.sh
@@ -6,8 +6,12 @@ RED="\033[38;5;196m"
 
 ## Collect packages to install and URL to download from 
 
+CROWNSTONE_SDK_TOOLS=https://github.com/crownstone/crownstone-sdk/releases/download/tools-0.0.1
+
+# Original website:
+# NORDIC_URL=https://developer.nordicsemi.com/nRF5_SDK/nRF5_SDK_v11.x.x/$NORDIC_SDK
 NORDIC_SDK=nRF5_SDK_11.0.0_89a8197.zip
-NORDIC_URL=https://developer.nordicsemi.com/nRF5_SDK/nRF5_SDK_v11.x.x/$NORDIC_SDK
+NORDIC_URL=$CROWNSTONE_SDK_TOOLS/$NORDIC_SDK
 
 # See for overview https://www.segger.com/downloads/jlink
 JLINK32=JLink_Linux_V610l_i386.deb
@@ -15,18 +19,27 @@ JLINK64=JLink_Linux_V610l_x86_64.deb
 JLINK32_URL=https://www.segger.com/downloads/jlink/$JLINK32
 JLINK64_URL=https://www.segger.com/downloads/jlink/$JLINK64
 
-GCC=gcc-arm-none-eabi-5_4-2016q2-20160622-linux.tar.bz2
-GCC_URL=https://launchpad.net/gcc-arm-embedded/5.0/5-2016-q2-update/+download/$GCC
+# Original website:
+# GCC_URL=https://launchpad.net/gcc-arm-embedded/5.0/5-2016-q2-update/+download/$GCC
+# Now at https://developer.arm.com/open-source/gnu-toolchain/gnu-rm/downloads
+GCC=gcc-arm-none-eabi-6_2-2016q4-20161216-linux.tar.bz2
+GCC_URL=$CROWNSTONE_SDK_TOOLS/$GCC
 
-# print function, first parameter is the string, second parameter the colour.
-# colour is optional
+# Print function:
+#   @param string
+#   @param colour (optional)
+#   @param --no-newline (optional)
 function print {
-    if [ -z $2 ]; then
-        echo -e $1
-    else
-        COLOR_RESET="\033[39m"
-        echo -e $2$1$COLOR_RESET
-    fi
+  args=""
+  if [ "$3" == "--no-newline" ]; then
+    args="-n"
+  fi
+  if [ -z $2 ]; then
+    echo $args -e $1
+  else
+    COLOR_RESET="\033[39m"
+    echo $args -e $2$1$COLOR_RESET
+  fi
 }
 
 print "> Checking architecture" $ORANGE
@@ -42,16 +55,21 @@ else
   exit 1
 fi
 
-print "> Installing script dependencies" $ORANGE
+print "> Install script dependencies with apt-get such as unzip, git and cmake [Y/n]: " $ORANGE --no-newline
 
-sudo apt-get install unzip git cmake -y
+read response
 
-if [[ $? -ne 0 ]]; then
-  print "Failed to install script dependencies" $RED
-  exit 1
+if [ "$response" == "n" ] || [ "$response" == "N" ]; then
+  print "Skip installing script dependencies" $RED
+else
+  sudo apt-get install unzip git cmake -y
+
+  if [[ $? -ne 0 ]]; then
+    print "Failed to install script dependencies" $RED
+    exit 1
+  fi
+  print "> OK" $GREEN
 fi
-
-print "> OK" $GREEN
 
 print "> Creating directories ..." $ORANGE
 
@@ -62,13 +80,13 @@ mkdir -p config
 mkdir -p nordic/v11/
 mkdir -p tools/compiler
 
-mkdir -p tmp
+mkdir -p downloads
 
 print "> OK" $GREEN
 
 print "> Getting Nordic SDK" $ORANGE
 
-cd tmp
+cd downloads
 wget $NORDIC_URL
 
 if [[ $? -ne 0 ]]; then
@@ -103,7 +121,7 @@ print ">> OK" $GREEN
 
 print "> Getting JLink" $ORANGE
 
-cd tmp
+cd downloads
 
 if $arch64; then
   echo "wget --post-data \"accept_license_agreement=accepted\" $JLINK64_URL"
@@ -119,26 +137,30 @@ fi
 
 print "> OK" $GREEN
 
-print ">> Installing JLink" $ORANGE
+print ">> Install JLink [Y/n]: " $ORANGE --no-newline
+read response
 
-if $arch64; then
-  sudo dpkg -i $JLINK64
+if [ "$response" == "n" ] || [ "$response" == "N" ]; then
+  print "Skip installing JLink" $RED
 else
-  sudo dpkg -i $JLINK32
-fi
+  if $arch64; then
+    sudo dpkg -i $JLINK64
+  else
+    sudo dpkg -i $JLINK32
+  fi
 
-if [[ $? -ne 0 ]]; then
-  print "Failed to install JLink" $RED
-  exit 1
+  if [[ $? -ne 0 ]]; then
+    print "Failed to install JLink" $RED
+    exit 1
+  fi
+  print ">> OK" $GREEN
 fi
 
 cd ..
 
-print ">> OK" $GREEN
-
 print "> Getting compiler" $ORANGE
 
-cd tmp
+cd downloads
 wget $GCC_URL
 
 if [[ $? -ne 0 ]]; then
@@ -159,24 +181,29 @@ fi
 
 print ">> OK" $GREEN
 
-print ">> Installing compiler dependencies" $ORANGE
+print ">> Install compiler dependencies (Y/n): " $ORANGE --no-newline
+read response
 
-sudo dpkg --add-architecture i386
-sudo apt-get update
-sudo apt-get install libstdc++6:i386 libncurses5:i386 -y
+if [ "$response" == "n" ] || [ "$response" == "N" ]; then
+  print "Skip installing JLink" $RED
+else
+  sudo dpkg --add-architecture i386
+  sudo apt-get update
+  sudo apt-get install libstdc++6:i386 libncurses5:i386 -y
 
-if [[ $? -ne 0 ]]; then
-  print "Failed to install compiler dependencies" $RED
-  exit 1
+  if [[ $? -ne 0 ]]; then
+    print "Failed to install compiler dependencies" $RED
+    exit 1
+  fi
+  print ">> OK" $GREEN
 fi
 
 cd ..
 
-print ">> OK" $GREEN
 
 print "> Getting bluenet" $ORANGE
 
-git clone https://github.com/crownstone/bluenet --branch sdk_11
+git clone https://github.com/crownstone/bluenet
 
 if [[ $? -ne 0 ]]; then
   print "Failed to clone into bluenet" $RED
@@ -215,13 +242,20 @@ echo "NRF51822_DIR=$PWD/nordic/v11" >> $PWD/config/CMakeBuild.config
 
 print "Please fill in the missing variables in the CMakeBuild.config (should be opened automatically in a gedit window). The script will complete once the window is closed."
 
-gedit $PWD/config/CMakeBuild.config
+xdg-open $PWD/config/CMakeBuild.config
 
 print ">> OK" $GREEN
 
 print "> Clean up" $ORANGE
 
-rm -r tmp
+print ">> Remove downloaded files (Y/n): " $ORANGE --no-newline
+read response
+
+if [ "$response" == "n" ] || [ "$response" == "N" ]; then
+  print "Skip deleting downloads" $GREEN
+else
+  rm -r downloads
+fi
 
 print "**************************************************************************" $GREEN
 print "***************************** DONE ***************************************" $GREEN
